Revert global changes, rename paths to be correct
