+++ src/renderer/components/+user-management/+cluster-role-bindings/details.scss
