+++ src/renderer/components/+user-management/+role-bindings/details.scss
