+++ src/renderer/components/app.scss
@@ -29,7 +29,7 @@
   --unit: 8px;
   --padding: var(--unit);
   --margin: var(--unit);
-  --border-radius: 5px;
+  --border-radius: 3px;
   --font-main: 'Roboto', 'Helvetica', 'Arial', sans-serif;
   --font-monospace: Lucida Console, Monaco, Consolas, monospace;
   --font-size-small: calc(1.5 * var(--unit));
